+++ src/But/Log/Destination/Console.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <iostream>
 #include "Stream.hpp"
-#include "Foregin.hpp"
+#include "Foreign.hpp"
 
 namespace But
 {
@@ -10,7 +10,7 @@ namespace Log
 namespace Destination
 {
 
-class Console final: public Foregin
+class Console final: public Foreign
 {
 public:
   Console() = default;
@@ -24,7 +24,7 @@ public:
   auto operator->() { return this; }
 
 private:
-  void logImpl(Backend::Entry e) override { static_cast<Foregin&>(s_).log( std::move(e) ); }
+  void logImpl(Backend::Entry e) override { static_cast<Foreign&>(s_).log( std::move(e) ); }
   void reloadImpl() override { s_.reload(); }
   void flushImpl() override { s_.flush(); }