+++ src/But/Log/Destination/ForeignMock.ut.hpp
@@ -1,9 +1,9 @@
 #include "gtest/gtest.h"
 #include "gmock/gmock.h"
-#include "Foregin.hpp"
+#include "Foreign.hpp"
 
 using testing::_;
-using But::Log::Destination::Foregin;
+using But::Log::Destination::Foreign;
 using But::Log::Backend::Entry;
 using But::Log::Backend::FieldInfo;
 
@@ -14,7 +14,7 @@ namespace Log
 namespace Destination
 {
 
-struct ForeginMock: public Foregin
+struct ForeignMock: public Foreign
 {
   MOCK_METHOD1(logImpl, void(Entry));
   MOCK_METHOD0(reloadImpl, void());