+++ src/But/Log/Destination/Multi.hpp
@@ -13,11 +13,11 @@ namespace Destination
 
 /** @brief template that can hold multiple destinations and forwards all calls to all of them.
  *
- *  @note this destination does NOT offer Foregin interface, since it is mean to be root element.
- *        in order to have multiple foregin interfaces, use MultiForegin base class.
+ *  @note this destination does NOT offer Foreign interface, since it is mean to be root element.
+ *        in order to have multiple foreign interfaces, use MultiForeign base class.
  *
- *  @note it is possible to have multiple destinations and foregin destinations, by creating
- *        set of destinations and one MultiForegin, that aggregates all foregin destinations, required.
+ *  @note it is possible to have multiple destinations and foreign destinations, by creating
+ *        set of destinations and one MultiForeign, that aggregates all foreign destinations, required.
  */
 template<typename ...Dsts>
 class Multi final