+++ src/But/Log/Destination/MultiForeign.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <vector>
 #include "But/NotNull.hpp"
-#include "Foregin.hpp"
+#include "Foreign.hpp"
 
 namespace But
 {
@@ -10,16 +10,16 @@ namespace Log
 namespace Destination
 {
 
-/** @brief template that can hold multiple foregin destinations and forwards all calls to all of them.
+/** @brief template that can hold multiple foreign destinations and forwards all calls to all of them.
  *  @note interface allows to change destinations at run time, eg. via config file.
  */
-class MultiForegin final: public Foregin
+class MultiForeign final: public Foreign
 {
 public:
-  using value_type = NotNullShared<Foregin>;
+  using value_type = NotNullShared<Foreign>;
   using collection_type = std::vector<value_type>;
 
-  explicit MultiForegin(collection_type dsts): dsts_{ std::move(dsts) } { }
+  explicit MultiForeign(collection_type dsts): dsts_{ std::move(dsts) } { }
 
 private:
   void logImpl(Backend::Entry e) override;