+++ src/But/Log/Destination/Syslog.manual.cpp
@@ -17,9 +17,9 @@ int main()
   s.log( Priority::warning, " this is warning");
   s.log( Priority::error, " this is error");
 
-  But::Log::Destination::Foregin& base = s;
-  base.log( Timestamp{}, " ", Priority::info, " @ ", FileName{__FILE__}, ":", LineNumber{__LINE__}, " ", "hello, foregin world" );
-  base.log( Timestamp{}, " ", Priority::info, " @ ", FileName{__FILE__}, ":", LineNumber{__LINE__}, " ", "bye, foregin world" );
+  But::Log::Destination::Foreign& base = s;
+  base.log( Timestamp{}, " ", Priority::info, " @ ", FileName{__FILE__}, ":", LineNumber{__LINE__}, " ", "hello, foreign world" );
+  base.log( Timestamp{}, " ", Priority::info, " @ ", FileName{__FILE__}, ":", LineNumber{__LINE__}, " ", "bye, foreign world" );
 
   base.log( Priority::debug, " this is debug");
   base.log( Priority::info, " this is info");