+++ src/But/Log/LoggerProxy.hpp
@@ -6,7 +6,7 @@ namespace Log
 {
 
 /** @brief proxy class, to be used as an entrypoint for logging.
- *         Destination type can be anything that derives from Destination::Foregin class,
+ *         Destination type can be anything that derives from Destination::Foreign class,
  *         or a user-provided type, that will handle all the arguments on its own.
  *
  * @example basic, zero-overhead, on-console output logger can be achieved like this:
@@ -16,13 +16,13 @@ namespace Log
  * log.log( Timestamp{}, " hello - current UTC date is: ", UtcDate{} );
  * </code>
  *
- * @example if a dynamic output is needed, use Destination::Foregin as a base class,
+ * @example if a dynamic output is needed, use Destination::Foreign as a base class,
  *          for implementing any output for you want. as a parameter you will receive a
  *          collection of type-value pairs, reprenseting each argument.
  *
  * @note this class offers an option to either use your Destination directly (eg. Destination::Console),
  *       so that you can configure logger with a sinle typedef, or provide a runtime-dependent
- *       argument (i.e. Destination::Foregin derived class) and allow to change destination, depending
+ *       argument (i.e. Destination::Foreign derived class) and allow to change destination, depending
  *       on eg. config file entries.
  *
  * @warning it is "destination"'s implemented responsability to handile input/output in a thread-safe manner!